Force asset MD5 to recalculate on save
This change blanks the MD5 for an asset once it is successfully loaded.
Since the file name of the asset contains the MD5 information, it's
possible to modify a file's contents without updating its filename or to
rename a file without changing its contents, causing the claimed MD5 to
be inaccurate. Previously this would trigger an error at save time; now
the client will recalculate the MD5 and everything should go smoothly.