Don't dupe shaders for stamping; always use PMA.
By setting the blend factors to (One,OneMinusSourceAlpha) at all times
we can use the same set of shaders both for rendering individual
costumes to bitmap and for rendering costumes on the stage. This
requires using pre-multiplied alpha (PMA) in the shader output, but that
only costs one instruction. In exchange we get correct blending for
pen-layer stamps and we never need to change blending modes for any
reason.