Bring back blend argument for drawChild()
I had removed the argument since pre-multiplied alpha lets us use the
same blend mode for both stamps and normal drawing, making for simpler
code. However, blending is cheaper when the destination factor is zero.
With that in mind it may be a better idea to turn off blending when we
can, especially on mobile platforms.