Bring back texture filtering workaround
The recent 3D refactor broke the `forcePixelate` flag, which used to
force the pixelate effect on under certain conditions. Using the
"touching color _?" block with thin lines depended on this feature:
using linear filtering often causes the thin line to be blended too much
with surrounding elements, resulting in failed detection.

This change is similar in spirit to the `forcePixelate` flag from
before, but uses texture filtering instead of the pixelate effect. This
should result in improved performance when the pixelate effect is not in
use.

This fixes LLK/scratch-flash#836