Apply video transparency as ghost effect
The previous method of making the video image transparent relied on the
fact that all objects were drawn with the same shader, and allowed the
video image's alpha value to be used in place of the ghost effect. This
broke when we started determining the shader on a per-object basis. Now
video transparency is implemented by applying the standard ghost effect
to the bitmap object representing the video image.

This fixes LLK/scratch-flash#837 and LLK/scratch-flash#838
See issues for test cases