Cleaned up Ant build script and added commit ID
This change makes SCRATCH::revision available as a string in the AS3
code. When built through IDEA this string will be 'dev' and when built
through Ant this string will contain information about the current git
commit. The information shown is currently the output of the command:
`git describe --dirty --tags --always`

The new file `build-common.xml` contains a macro that does the work to
collect that information. This macro should be usable from the online
and offline editor build scripts.