First pass at logging system for the editor
The idea with this logging system is that we will record up to a fixed
number of events, and if a major error occurs then we will attach the
contents of the log (the last N events) to the error report. This
logging system tries to optimize its GC impact for that pattern of use:
in particular, logging events should have minimal GC impact -- maybe
even none if the extraData field isn't used and messageKey strings are
reused. Attaching the log to an error report will have significant GC
impact, but ideally that should be rare.