Partial fix for renaming cloud variables
When putting a variable's current name into the text input field for
renaming that variable, this change omits the cloud symbol. Since we
prepend a cloud symbol when renaming a cloud variable, that means
accepting the suggested name should be a no-op. Prior to this change,
doing so would duplicate the cloud symbol.
There is still some incorrect behavior going on when renaming a cloud
variable, which I suspect is related to server communication.