Change the way paint actions are committed on undo
On undo or redo, we now reactivate the tool that currently looks like
it's active. This means that pressing undo with an action in progress
commits that action just like changing a tool would, reducing the number
of code paths which commit an action. The old way was not working if the
undo button was pressed immediately after drawing an ellipse, a
rectangle, or some text.
This fixes #678