Open internal extension tips with 'ext:' tag
Internal extensions have special tips content formatted for the built-in
tips window. These pages are accessed by passing a specially formatted
tag to the tips API in JavaScript -- for example 'ext:LEGO WeDo' for the
WeDo extension. This change makes the editor supply that sort of tag
when opening tips for an internal extension. The behavior is unchanged
for non-internal extensions.

This fixes LLK/scratchr2#2709, assuming the tag -> URL mapping is
correct on the JavaScript side of the tips system.