Fix `arrayToBlock()` for null substacks in RTL
When the user selects a new language, the process of updating the script
blocks with new strings involves converting them to an array-based
representation and back. The way this process was distinguishing between
block arguments and substacks had trouble with `null` values, leading to
situations where substacks could be eliminated after being interpreted
as extraneous block arguments. The new code uses the number of block
arguments to avoid the need for such inspection.
This fixes LLK/scratch-flash#942