Reset requester state when args might have changed
The `collectArgs()` function runs when anything changes in the arguments
of a block or any of its descendants, recursively. If this is called on
a Requester block which currently has a value, reset / discard that
value in case it's no longer valid given the argument change.
This resolves LLK/scratchx#61