Use U+001F to separate extension name from op
Prior to the WeDo 2.0 extension, Scratch assumed that an extension's
name would never contain `.` -- Scratch would split an extension op into
an extension name and an op within that extension by looking for the
first `.`.

With the WeDo 2.0 extension, I changed the code to split on the last
`.`. Unfortunately this broke extensions which use `.` in their blocks.

This change uses U+001F (Unicode "Unit Separator") as the separator
instead of `.` since U+001F is very unlikely to be contained in an
extension's name or blocks. The separator and unpacking logic have been
centralized in `ExtensionManager` as well.

This resolves #1044