Handle image load errors more gracefully
Prior to this change, any error encountered on any image (costume or
backdrop) in the project would abort the `decodeAllImages` process.
Unfortunately this was handled in such a way that it would cause the
project to never finish loading, resulting in projects getting stuck
at "100%" on the loading progress bar.
Now, images which fail to load are replaced with an empty bitmap
costume. In addition, such failures are reported to the user in a dialog
at the end of the loading process. This report includes the sprite name
(if applicable) and the costume/backdrop name for each failure.