Check more carefully for Scratch stage clicks
Before: any click within the rectangle of the Scratch stage would be
forwarded to the top-most sprite at that (x,y) location, if any, even if
the sprite was covered by some other UI such as a media library window.

After: a click is forwarded to the top-most sprite at its (x,y) location
if Flash reports the target of the click is either the Flash stage (this
covers 3D mode) or some descendant of the Scratch stage (this covers 2D
mode).