Separate presentation mode request & reaction
For better or worse, the editor tracks "presentation mode" (full-screen)
state in the state of the presentation mode toggle button. This is fine
as long as that button is the only way to change presentation mode, but
sometimes we want JavaScript to set the presentation mode state. This
change separates the concept of requesting a mode change (such as when
JS requests a state) from the reaction to a change (such as when the
button is clicked, or the JS request results in a change).