Ensure MD5 always matches soundData, add logging
The `md5` of a `ScratchSound` object is now a read-only getter which
calculates the MD5 if the `soundData` has changed since last time the
`md5` was accessed. There should no longer be any way to get the MD5 out
of sync with the actual samples.

I also added warning messages to several of the paths in `ScratchSound`
which can result in empty sounds or samples.