Fix Boolean reporters in extensions when offline
Extension reporters must run asynchronously in the offline editor. This
change makes Boolean reporters act like requesters in the offline
editor, just like we do with non-Boolean reporters.

Note that we don't normally support Boolean requesters -- there's no
`'B'` block type -- but the code to handle extension requesters doesn't
care whether the block is Boolean or not so this change is sufficient.
It would be trivial to add support for `'B'` now.