Fix loading sounds by MD5
During project load, there's a brief time during which a `ScratchSound`
needs an MD5 corresponding to its upcoming contents -- not its current
contents. This precludes the MD5 field being read-only. This change
restores a line of code which sets the sound's MD5 from loaded JSON.

Additionally, the sound ID is now always set to -1 in `prepareToSave`.
This ensures that the sound will be reallocated a slot in the list of
saved sounds. Even if this sound has not changed another sound may have
been added or removed, which could lead to this sound's old ID pointing
to the wrong entry in the array.

Note that the list of saved sounds is rebuilt each save, so there's no
need to deallocate the entry corresponding to the sound's old ID. See
`recordImagesAndSounds` in `ProjectIO.as` for more details.