Catch and log malformed WAV errors
If `WAVFile.extractSamples()` is called with malformed (including empty)
`waveData` then `WAVFile.decode` will throw. This can cause the Sounds
tab in the editor to fail initialization, resulting in a blank white
area instead of the Sounds tab. This change catches and logs such an
exception, returning an "empty" (one-sample) result instead.