Enforce bounds in ListWatcher's setWidthHeight()
The `ListWatcher` class creates a cell for each visible item as part of
its resize handling (see `updateContents`). This can take a long time if
the height of the list watcher and the length of its list are both very
large. This is especially true at load-time, when a list watcher may not
yet know what its true size should be.

This change enforces bounds in the `setWidthHeight()` method of the
`ListWatcher` class. Bounding the size effectively bounds the number of
list cells that will be created, eliminating long load-time delays.