Skip POSTs with empty body, output a warning
Flash doesn't correctly handle a POST with an empty body: the URL
handler automatically converts the POST into a GET. This causes a 403
leading the editor to think the user has logged out. The end result is a
save failure that's to recover from.

A project shouldn't have a zero-length asset anyway, but if it does get
into that situation then this change will prevent the entire project
from being "lost" due to a single-asset error.