Don't list English twice in the language selector
Previous behavior: Always list English first, then append the languages
retrieved from the server. If English is also in that list, then English
will be listed twice.

New behavior: Use the language list retrieved from the server exactly.
If we fail to retrieve a list from the server, use English as a
fallback.