Translation: ask extensions for their specs
This change removes the hard-coded list of extension specs from
`Specs.as`, replacing it with a function on `ExtensionManager` which
reads block specs from the actual extensions. This requires the
extensions to be loaded, so the function pops up a warning dialog if it
detects that any built-in extension is not loaded.