Restore script position when another script is selected
It also restores the script postiion when saving.

Another thing that would be nice is to have the top block
undraggable (when only one script is viewed), since it gets
restored back to its old position.. or would that be bad?

It's not the most ideal option, but I think it's important
enough to have fixed like this for now.

In the future I'd like a way to clean up blocks akin to
this: https://scratch.mit.edu/projects/95417277/

..But I'm not really sure how to do that, honestly.

@Zro617 can you ping people in commit messages??????