Warn about missing extension specs only on export
The new `getExtensionSpecs` method can pop up a warning if one or more
built-in extensions are not loaded. This helps avoid problems when
exporting strings for translation, but we don't want this warning to pop
up when users change their language setting. This change introduces a
Boolean argument to control whether or not the warning should pop up.