Collect field text from numeric fields only
This change preserves user-entered text for a block argument, but only
in the specific case of a numeric block argument where the string
representation of the numeric value doesn't match the field text. This
means that if the user enters "42" then the serialized value will be a
numeric 42, but if the user enters "42." then the serialized value will
be that string exactly. Other argument types should not be affected by
this change.