Use keyCode instead of charCode for primKeyPressed
The `keyIsDown` array now tracks keys by their normalized `keyCode`
rather than ASCII value. This means that in situations where we don't
get a good `charCode` (see #1334) we can still track a key without
worrying about mapping its `keyCode` to an ASCII value. That mapping is
straightforward for some keys (letters, for example) but not for others
(comma, for example).