Revert Edge keyboard fixes, keeping some cleanup
Edge has fixed their `charCode` bug, so we don't need to work around it
any more. However, I felt that some of the changes I made for the
workaround were worth keeping. This code should have the same observable
behavior as the pre-Edge-fix code, but logic related to keyboard events
and key codes & names is centralized and better documented. Hopefully
this will make this a more useful reference as we run into edge cases
and quirks in Scratch 3.0...